Microsoft onedrive fixes
Looks like lots poeple having trouble. But our ttemp to fix is in this change,
not sure if the mssfix does take effect or the removing def1 gateway or stop
pushing the dns server, but I dont want to test anymore so leave it like this.

It does not affect performance or anything, just that if we need to push dns
server we might need to re-test it again and for MS client we might filter out
the pushing.

I suspect removing the def1 makes it works - as per this
https://www.macwheeler.com/windows-10-office-365-cannot-connect-over-openvpn-fixed/